use separated launcher for MergeJars
