+++ separated/src/main/java/cpw/mods/fml/relauncher/Side.java
@@ -1,0 +1,41 @@
+/*
+ * Forge Mod Loader
+ * Copyright (c) 2012-2013 cpw.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the GNU Lesser Public License v2.1
+ * which accompanies this distribution, and is available at
+ * http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
+ *
+ * Contributors:
+ *     cpw - implementation
+ */
+
+package cpw.mods.fml.relauncher;
+
+public enum Side {
+
+    /**
+     * The client side. Specifically, an environment where rendering capability exists.
+     * Usually in the game client.
+     */
+    CLIENT,
+    /**
+     * The server side. Specifically, an environment where NO rendering capability exists.
+     * Usually on the dedicated server.
+     */
+    SERVER;
+
+    /**
+     * @return If this is the server environment
+     */
+    public boolean isServer() {
+        return !isClient();
+    }
+
+    /**
+     * @return if this is the Client environment
+     */
+    public boolean isClient() {
+        return this == CLIENT;
+    }
+}
